wip: Move OpenLayers map and utilities out of React components
